backend/requirements.txt
alembic==1.16.5
annotated-types==0.7.0   
anyio==4.10.0
APScheduler==3.11.0      
attrs==25.3.0
bcrypt==3.2.2
beautifulsoup4==4.13.5   
certifi==2025.8.3        
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudscraper==1.2.71     
colorama==0.4.6
curl_cffi==0.13.0        
dnspython==2.8.0
email-validator==2.3.0   
fastapi==0.117.1
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
lxml==6.0.1
Mako==1.3.10
MarkupSafe==3.0.2
outcome==1.3.0.post0
passlib==1.7.4
playwright==1.55.0
psycopg==3.2.10
psycopg-binary==3.2.10
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyee==13.0.0
PyJWT==2.10.1
pyparsing==3.2.4
PySocks==1.7.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
selenium==4.35.0
setuptools==80.9.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.43
sse-starlette==3.0.2
starlette==0.48.0
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
undetected-chromedriver==3.5.5
urllib3==2.5.0
uvicorn==0.36.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0